New param for FollowRedirect Chromium change: https://chromium.googlesource.com/chromium/src/+/eb2ba20c58a0abfd5585707ad57d67c49dcb397d OOR-CORS: Use cors_exempt_headers for X-GoogApps-Allowed-Domains
This patch converts the required header handling for
X-GoogApps-Allowed-Domains to use cors_exempt_headers.

To realize it, we need to modify following existing interfaces to
allow touching cors_exempt_headers on redirects.

- blink::URLLoaderThrottle::WillRedirectRequest
- network::mojom::URLLoader::FollowRedirect

With these update, GoogleURLLoaderThrottle handles
X-GoogApps-Allowed-Domains header as a member of cors_exempt_headers
even on redirects, and this allow the header to be exempted from
CORS checks.

CorsURLLoader validates the modified cors_exempt_headers in the
FollowRedirect, and network::URLLoader merge it into the request
headers for //net API.

SystemNetworkContextManager registers the X-GoogApps-Allowed-Domains
header to the cors_exempt_header allowlist for the secondary security
check in the CorsURLLoader as explained above.

Bug: 1067166
Change-Id: I9f29a5ef17e68180803b867ec817946f5d3510e1
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/2129787
Commit-Queue: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Scott Violet <sky@chromium.org>
Reviewed-by: Karan Bhatia <karandeepb@chromium.org>
Reviewed-by: Sergey Volk <servolk@chromium.org>
Reviewed-by: Jochen Eisinger <jochen@chromium.org>
Reviewed-by: Richard Coles <torne@chromium.org>
Reviewed-by: Kinuko Yasuda <kinuko@chromium.org>
Reviewed-by: Wez <wez@chromium.org>
Reviewed-by: Sami Kyöstilä <skyostil@chromium.org>
Cr-Commit-Position: refs/heads/master@{#762882}